Cmd类封装，增加析构函数、终止进程方法、数据接收事件

+++ DataSmith/Core/Util/Cmd.cs
@@ -4,26 +4,49 @@ using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Text;
+using Quartz.Util;
 
 namespace DataSmith.Core.Util
 {
     public class Cmd
     {
-        private readonly Process _process = new Process();
-        public DataReceivedEventHandler DataReceivedHandler { get; set; }
+        private Process _process;
+
+        /// <summary>
+        ///   每次应用程序向其重定向 <see cref="P:System.Diagnostics.Process.StandardOutput" /> 流中写入行时发生。
+        /// </summary>
+        [Browsable(true)]
+        [MonitoringDescription("ProcessAssociated")]
+        public event DataReceivedEventHandler OutputDataReceived;
 
         public Cmd()
         {
+            Init();
+        }
+
+        ~Cmd()
+        {
+            Abort();
+        }
+
+        private void Init()
+        {
+            if (_process != null)
+                return;
+
+            _process = new Process();
             _process.OutputDataReceived -= new DataReceivedEventHandler(ProcessOutDataReceived);
-            ProcessStartInfo p = new ProcessStartInfo();
-            p.FileName = "cmd.exe";
-            //p.Arguments = " -t 192.168.1.103";
-            p.UseShellExecute = false;
-            p.WindowStyle = ProcessWindowStyle.Hidden;
-            p.CreateNoWindow = true;
-            p.RedirectStandardError = true;
-            p.RedirectStandardInput = true;
-            p.RedirectStandardOutput = true;
+            ProcessStartInfo p = new ProcessStartInfo
+            {
+                FileName = "cmd.exe",
+                //p.Arguments = " -t 192.168.1.103";
+                UseShellExecute = false,
+                WindowStyle = ProcessWindowStyle.Hidden,
+                CreateNoWindow = true,
+                RedirectStandardError = true,
+                RedirectStandardInput = true,
+                RedirectStandardOutput = true
+            };
             _process.StartInfo = p;
             _process.Start();
 
@@ -33,13 +56,20 @@ namespace DataSmith.Core.Util
 
         public void Exec(string commond, string parameter = "")
         {
+            if (_process == null)
+                Init();
+
             _process.StandardInput.WriteLine($"{commond} {parameter}");
         }
 
-        public void Exit()
+        public void Abort()
         {
-            _process.CancelOutputRead();//取消异步操作
-            _process.Kill();
+            if (_process != null)
+            {
+                _process.CancelOutputRead();//取消异步操作
+                _process.Kill();
+                _process = null;
+            }
         }
 
         /// <summary>
@@ -47,9 +77,34 @@ namespace DataSmith.Core.Util
         /// </summary>
         /// <param name="sender"></param>
         /// <param name="e"></param>
-        public void ProcessOutDataReceived(object sender, DataReceivedEventArgs e)
+        private void ProcessOutDataReceived(object sender, DataReceivedEventArgs e)
         {
-            DataReceivedHandler?.Invoke(sender, e);
+            OutputDataReceived?.Invoke(sender, e);
         }
+
+        ///// <summary>
+        ///// 进程接受事件
+        ///// </summary>
+        ///// <param name="sender"></param>
+        ///// <param name="e"></param>
+        //public void ProcessOutDataReceived(object sender, DataReceivedEventArgs e)
+        //{
+        //    if (txtOutPutInfo.InvokeRequired)
+        //    {
+        //        txtOutPutInfo.Invoke(new Action(() =>
+        //        {
+        //            txtOutPutInfo.AppendText(e.Data + "\r\n");
+        //            //txtOutPutInfo.SelectionStart = txtOutPutInfo.Text.Length;
+        //            //txtOutPutInfo.SelectionLength = 0;
+        //            //txtOutPutInfo.Focus();
+        //            txtOutPutInfo.ScrollToCaret();
+        //        }));
+        //    }
+        //    else
+        //    {
+        //        txtOutPutInfo.AppendText(e.Data + "\r\n");
+        //    }
+        //}
+
     }
 }
